+++ baseapp/abci.go
@@ -68,7 +68,13 @@ func (app *BaseApp) InitChain(req *abci.RequestInitChain) (*abci.ResponseInitCha
 	// Store the consensus params in the BaseApp's param store. Note, this must be
 	// done after the finalizeBlockState and context have been set as it's persisted
 	// to state.
-	if req.ConsensusParams != nil {
+	if cp := req.ConsensusParams; cp != nil {
+		if cp.Version == nil || (cp.Version != nil && cp.Version.App == 0) {
+			cp.Version = &cmtproto.VersionParams{
+				App: InitialAppVersion,
+			}
+		}
+
 		err := app.StoreConsensusParams(app.finalizeBlockState.Context(), *req.ConsensusParams)
 		if err != nil {
 			return nil, err
@@ -136,11 +142,22 @@ func (app *BaseApp) InitChain(req *abci.RequestInitChain) (*abci.ResponseInitCha
 
 func (app *BaseApp) Info(_ *abci.RequestInfo) (*abci.ResponseInfo, error) {
 	lastCommitID := app.cms.LastCommitID()
+	appVersion := InitialAppVersion
+	if lastCommitID.Version > 0 {
+		ctx, err := app.CreateQueryContext(lastCommitID.Version, false)
+		if err != nil {
+			return nil, fmt.Errorf("failed creating query context: %w", err)
+		}
+		appVersion, err = app.AppVersion(ctx)
+		if err != nil {
+			return nil, fmt.Errorf("failed getting app version: %w", err)
+		}
+	}
 
 	return &abci.ResponseInfo{
 		Data:             app.name,
 		Version:          app.version,
-		AppVersion:       app.appVersion,
+		AppVersion:       appVersion,
 		LastBlockHeight:  lastCommitID.Version,
 		LastBlockAppHash: lastCommitID.Hash,
 	}, nil
@@ -192,7 +209,8 @@ func (app *BaseApp) Query(_ context.Context, req *abci.RequestQuery) (resp *abci
 
 	case QueryPathP2P:
 		resp = handleQueryP2P(app, path)
-
+	case QueryPathCustom:
+		resp = handleQueryCustom(app, path, req)
 	default:
 		resp = sdkerrors.QueryResult(errorsmod.Wrap(sdkerrors.ErrUnknownRequest, "unknown query path"), app.trace)
 	}
@@ -1087,6 +1105,43 @@ func handleQueryP2P(app *BaseApp, path []string) *abci.ResponseQuery {
 	}
 
 	return resp
+}
+
+func handleQueryCustom(app *BaseApp, path []string, req *abci.RequestQuery) *abci.ResponseQuery {
+	// path[0] should be "custom" because "/custom" prefix is required for keeper
+	// queries.
+	//
+	// The QueryRouter routes using path[1]. For example, in the path
+	// "custom/gov/proposal", QueryRouter routes using "gov".
+	if len(path) < 2 || path[1] == "" {
+		return sdkerrors.QueryResult(errorsmod.Wrap(sdkerrors.ErrUnknownRequest, "no route for custom query specified"), app.trace)
+	}
+
+	querier := app.customQueryRouter.Route(path[1])
+	if querier == nil {
+		return sdkerrors.QueryResult(errorsmod.Wrapf(sdkerrors.ErrUnknownRequest, "no custom querier found for route %s", path[1]), app.trace)
+	}
+
+	ctx, err := app.CreateQueryContext(req.Height, req.Prove)
+	if err != nil {
+		return sdkerrors.QueryResult(err, app.trace)
+	}
+
+	// Passes the rest of the path as an argument to the querier.
+	//
+	// For example, in the path "custom/gov/proposal/test", the gov querier gets
+	// []string{"proposal", "test"} as the path.
+	resBytes, err := querier(ctx, path[2:], req)
+	if err != nil {
+		res := sdkerrors.QueryResult(err, app.trace)
+		res.Height = req.Height
+		return res
+	}
+
+	return &abci.ResponseQuery{
+		Height: req.Height,
+		Value:  resBytes,
+	}
 }
 
 // SplitABCIQueryPath splits a string path using the delimiter '/'.